+++ src/graphql/entity/create_field/resolve_fields/resolve_document_field/mod.rs
@@ -20,7 +20,10 @@ impl ServiceEntity {
 
         match resolved {
             DocumentValue::ObjectID(object_id) => Ok(Value::from(object_id.to_string())),
-            _ => unreachable!("Invalid result type for object id scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for object id scalar",
+            )),
         }
     }
 
@@ -39,7 +42,10 @@ impl ServiceEntity {
             DocumentValue::StringArray(values) => Ok(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
             )),
-            _ => unreachable!("Invalid result type for string scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for string scalar",
+            )),
         }
     }
 
@@ -59,7 +65,10 @@ impl ServiceEntity {
             DocumentValue::IntArray(values) => Ok(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
             )),
-            _ => unreachable!("Invalid result type for int scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for int scalar",
+            )),
         }
     }
 
@@ -79,7 +88,10 @@ impl ServiceEntity {
             DocumentValue::BooleanArray(values) => Ok(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
             )),
-            _ => unreachable!("Invalid result type for boolean scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for boolean scalar",
+            )),
         }
     }
 
@@ -102,7 +114,10 @@ impl ServiceEntity {
                     .map(|value| Value::from(value.to_string()))
                     .collect(),
             )),
-            _ => unreachable!("Invalid result type for UUID scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for UUID scalar",
+            )),
         }
     }
 
@@ -125,7 +140,10 @@ impl ServiceEntity {
                     .map(|value| Value::from(value.to_rfc3339()))
                     .collect(),
             )),
-            _ => unreachable!("Invalid result type for DateTime scalar"),
+            DocumentValue::None => Ok(Value::Null),
+            _ => Err(async_graphql::Error::from(
+                "Invalid result type for DateTime scalar",
+            )),
         }
     }